+++ src/components/js/CardList.js
@@ -1,41 +1,6 @@
 import "../css/Card.css";
-import logoImg from "../../Assets/noImgLogo.svg";
-
-/* 게시물 올린 시간 함수 */
-function getTimeDiff(value) {
-  const milliSeconds = new Date() - value;
-  const seconds = milliSeconds / 1000;
-  if (seconds < 60) return `방금 전`;
-  const minutes = seconds / 60;
-  if (minutes < 60)
-    return `${Math.floor(minutes)}${
-      Math.floor(minutes) > 1 ? " minutes ago" : " minute ago"
-    }`;
-  const hours = minutes / 60;
-  if (hours < 24)
-    return `${Math.floor(hours)}${
-      Math.floor(hours) > 1 ? " hours ago" : " hour ago"
-    }`;
-  const days = hours / 24;
-  if (days < 7)
-    return `${Math.floor(days)}${
-      Math.floor(days) > 1 ? " days ago" : " day ago"
-    }`;
-  const weeks = days / 7;
-  if (weeks < 5)
-    return `${Math.floor(weeks)}${
-      Math.floor(weeks) > 1 ? " weeks ago" : " week ago"
-    }`;
-  const months = days / 30;
-  if (months < 12)
-    return `${Math.floor(months)}${
-      Math.floor(months) > 1 ? " months ago" : " month ago"
-    }`;
-  const years = days / 365;
-  return `${Math.floor(years)}${
-    Math.floor(years) > 1 ? " years ago" : " year ago"
-  }`;
-}
+import noImage from "../../Assets/noImage.png";
+import getTimeDiff from "../../utils/utilTimeDiff";
 
 /* 각 카드 컴포넌트 */
 function CardItem({ item, key }) {
@@ -45,14 +10,14 @@ function CardItem({ item, key }) {
   };
 
   const nowDate = getTimeDiff(new Date(item.createdAt));
-  const createdDate = item.createdAt.split("T")[0].split("-");
+  const createdDate = item.createdAt.split("T")[0].split("-").join(".");
 
   return (
     <a href={item.url}>
       <div key={key} className="card">
         <div className="card-img-wrap">
           {!item.imageSource ? (
-            <img className="logoImg" src={logoImg} alt={logoImg} />
+            <img className="logoImg" src={noImage} alt={noImage} />
           ) : (
             <div className="card-img" style={imgStyle}></div>
           )}
@@ -60,7 +25,7 @@ function CardItem({ item, key }) {
         <div className="card-information">
           <div className="time">{nowDate}</div>
           <p>{item.description}</p>
-          <div className="day">{createdDate.join(".")}</div>
+          <div className="day">{createdDate}</div>
         </div>
       </div>
     </a>
